style(AuthModule) : rename bearer.guard.ts
- make consistency file with class name
