+++ src/auth/auth.controller.ts
@@ -40,8 +40,8 @@ import { VerifyDTO } from './dto/verify.dto';
 import { OneTimeToken, OneTimeTokenPurpose } from './entity/one-time-token.entity';
 import { Verification } from './entity/verification.entity';
 import { BasicTokenGuard } from './guard/authentication/basic.guard';
-import { TokenAuthGuard } from './guard/authentication/bearer.guard';
 import { SecurityTokenGuard } from './guard/authentication/security-token.guard';
+import { TokenAuthGuard } from './guard/authentication/token-auth.guard';
 import { OwnerGuard } from './guard/authorization/owner.guard';
 import {
   AuthUserPayload,
