+++ src/auth/auth.module.ts
@@ -8,9 +8,9 @@ import { AuthService } from './auth.service';
 import { OneTimeToken } from './entity/one-time-token.entity';
 import { Verification } from './entity/verification.entity';
 import { BasicTokenGuard } from './guard/authentication/basic.guard';
-import { TokenAuthGuard } from './guard/authentication/bearer.guard';
 import { SecurityTokenGuard } from './guard/authentication/security-token.guard';
 import { TempUserGuard } from './guard/authentication/temp-user.guard';
+import { TokenAuthGuard } from './guard/authentication/token-auth.guard';
 import { OwnerGuard } from './guard/authorization/owner.guard';
 import { RolesGuard } from './guard/authorization/roles.guard';
